Moves AccountsDbConfig to its own submodule
